check for and prevent infinite recursion, cache qty in part info

+++ release.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 python manage.py test
-vim setup.py
-rm dist/*
-python setup.py sdist
-twine upload dist/*
+read -p "Continue? " -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    vim setup.py
+    rm dist/*
+    python setup.py sdist
+    twine upload dist/*
+fi
